# packages/api/swagger.json
{
  "openapi": "3.0.1",
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local Server"
    }
  ],
  "info": {
    "version": "0.0.1",
    "title": "nestjs",
    "description": "",
    "license": {
      "name": "UNLICENSED"
    }
  },
  "paths": {
...
    "/api/auth": {
      "get": {
        "tags": ["Auth"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SUCCESS__type"
                }
              }
            }
          },
          "400": {
            "description": "토큰이 유효하지 않음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError.Authentication.TOKEN_INVALID"
                }
              }
            }
          }
        },
        "summary": "로그인 여부 확인 API",
        "description": "로그인 여부 확인\n\n로그인 여부를 확인하는 API 입니다.\n\n## Error Case\n- TOKEN_INVALID: 토큰이 유효하지 않음\n- TOKEN_MISSING: 토큰이 없음",
        "security": [
          {
            "access_token": []
          }
        ]
      }
    },
    "/api/auth/token": {
      "get": {
        "tags": ["Auth"],
        "parameters": [
          {
            "name": "X-Client-Type",
            "in": "header",
            "schema": {
              "type": "string",
              "default": "'WEB'",
              "description": "Client type"
            },
            "description": "Client type",
            "required": false
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SUCCESSaccess_tokenstringrefresh_tokenstringaccess_tokenstring"
                }
              }
            }
          },
          "400": {
            "description": "토큰이 없음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError.Authentication.TOKEN_MISSING"
                }
              }
            }
          },
          "401": {
            "description": "토큰이 유효하지 않음",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError.Authentication.TOKEN_INVALID"
                }
              }
            }
          }
        },
        "summary": "refresh token api",
        "description": "refresh token\n\n리프레시 토큰을 이용하여 액세스 토큰을 재발급합니다.\n리프레시 토큰이 없거나 유효하지 않으면 에러를 반환합니다.\n\n**주의**\n앱의경우 헤더로, 웹의 경우 쿠키로 리프레시 토큰을 전달해야합니다.\n\n한번 사용된 리프레시 토큰은 더이상 사용할 수 없습니다.\n만약 리프레시 토큰을 재사용하면 해당 유저의 리프레시 토큰은 모두 만료됩니다.\n\n\n## Error Case\n- TOKEN_INVALID: 토큰이 유효하지 않음\n- TOKEN_MISSING: 토큰이 없음",
        "security": [
          {
            "refresh_token": []
          },
          {
            "web_refresh_token": []
          }
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "tags": ["Auth"],
        "parameters": [
          {
            "name": "X-Client-Type",
            "in": "header",
            "schema": {
              "type": "string",
              "default": "'WEB'",
              "description": "Client type"
            },
            "description": "Client type",
            "required": false
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/Auth.Oauth"
                  },
                  {
                    "$ref": "#/components/schemas/Auth.Local"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/SUCCESS__type.o2"
                    },
                    {
                      "$ref": "#/components/schemas/SUCCESS__type.o1"
                    },
                    {
                      "$ref": "#/components/schemas/AuthError.User.USER_NOT_FOUND"
                    },
                    {
                      "$ref": "#/components/schemas/AuthError.Authentication.INVALID_TYPE"
                    },
                    {
                      "$ref": "#/components/schemas/AuthError.Authentication.INVALID_PASSWORD"
                    },
                    {
                      "$ref": "#/components/schemas/AuthError.SocialAuth.SOCIAL_AUTH_FAILED"
                    },
                    {
                      "$ref": "#/components/schemas/AuthError.SocialAuth.SOCIAL_AUTH_INFO_MISSING"
                    },
                    {
                      "$ref": "#/components/schemas/AuthError.SocialAuth.SOCIAL_SERVICE_ACCESS_DENIED"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "type이 유효하지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError.Authentication.INVALID_TYPE"
                }
              }
            }
          },
          "401": {
            "description": "비밀번호가 일치하지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError.Authentication.INVALID_PASSWORD"
                }
              }
            }
          },
          "403": {
            "description": "소셜로그인 서비스 접근 거부",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError.SocialAuth.SOCIAL_SERVICE_ACCESS_DENIED"
                }
              }
            }
          },
          "404": {
            "description": "이메일이 존재하지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError.User.USER_NOT_FOUND"
                }
              }
            }
          }
        },
        "summary": "로그인 API",
        "description": "Login\n - 리프레시토큰 전략에 대한 설명\n\n\n소셜로그인과 로컬로그인을 모두 지원합니다.\n각 로그인방식에따라 body에 필요한 정보가 다릅니다.\n\n**주의**\n앱의 경우 헤더에 x-client-type을 명시 해야합니다.\nx-client-type이 명시되지 않으면 웹으로 간주합니다.\n\n리프레시 토큰 전달:\n- 웹의 경우 쿠키로 전달됩니다.\n- 앱의 경우 바디로 전달됩니다."
      }
    },
    "/api/auth/signup": {
      "post": {
        "tags": ["Auth"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/Auth.Oauth"
                  },
                  {
                    "$ref": "#/components/schemas/Localnicknamestringemail_certification_idstring"
                  }
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/SUCCESS__type.o3"
                    },
                    {
                      "$ref": "#/components/schemas/AuthError.Authentication.INVALID_TYPE"
                    },
                    {
                      "$ref": "#/components/schemas/AuthError.SocialAuth.SOCIAL_AUTH_FAILED"
                    },
                    {
                      "$ref": "#/components/schemas/AuthError.SocialAuth.SOCIAL_AUTH_INFO_MISSING"
                    },
                    {
                      "$ref": "#/components/schemas/AuthError.SocialAuth.SOCIAL_SERVICE_ACCESS_DENIED"
                    },
                    {
                      "$ref": "#/components/schemas/AuthError.User.EMAIL_ALREADY_EXISTS"
                    },
                    {
                      "$ref": "#/components/schemas/AuthError.User.NICKNAME_ALREADY_EXISTS"
                    },
                    {
                      "$ref": "#/components/schemas/AuthError.Authentication.EMAIL_CERTIFICATION_NOT_VERIFIED"
                    },
                    {
                      "$ref": "#/components/schemas/AuthError.SocialAuth.SOCIAL_ACCOUNT_LINKING_FAILED"
                    },
                    {
                      "$ref": "#/components/schemas/AuthError.SocialAuth.SOCIAL_SERVICE_RESPONSE_ERROR"
                    },
                    {
                      "$ref": "#/components/schemas/AuthError.SocialAuth.SOCIAL_ACCOUNT_ALREADY_LINKED"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "이메일이 이미 존재합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError.User.EMAIL_ALREADY_EXISTS"
                }
              }
            }
          },
          "401": {
            "description": "이메일 인증번호가 일치하지 않습니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError.Authentication.EMAIL_CERTIFICATION_NOT_VERIFIED"
                }
              }
            }
          },
          "403": {
            "description": "소셜로그인 서비스 접근 거부",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError.SocialAuth.SOCIAL_SERVICE_ACCESS_DENIED"
                }
              }
            }
          },
          "500": {
            "description": "소셜계정 연동 실패",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError.SocialAuth.SOCIAL_ACCOUNT_LINKING_FAILED"
                }
              }
            }
          }
        },
        "summary": "회원가입 API",
        "description": "회원가입 API\n\n소셜과 로컬을 모두 지원합니다.\n각 회원가입방식에따라 body에 필요한 정보가 다릅니다.\n\n로컬 회원가입시, 이메일 인증이 필요합니다.\n이메일 인증후, 이메일 인증ID를 body에 포함하여 회원가입을 진행합니다.\n\n회원가입을 할때, email을 기준으로 유저를 구분합니다.\n다만 소셜회원가입 이후, 특별한 이유로 인하여 Provider(로그인 제공업체)의 이메일이 변경될경우, 해당 이메일로 로그인시에도 기존의 유저정보를 사용합니다.\n이는 소셜로그인의 특성상, 이메일이 변경될수 있기 때문입니다.\n\n**주의**\n로컬회원가입 후 소셜회원가입을 할시 자동으로 소셜계정과 연동됩니다.\n하지만, 소셜회원가입 후 로컬회원가입을 할시, 이미 가입된 이메일이라는 에러가 발생합니다.\n해당 문제는 비밀번호 찾기 flow를 통해 해결할 수 있습니다."
      }
    },
    "/api/auth/email/certification": {
      "post": {
        "tags": ["Auth"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Auth.SendEmailVerificationCodeDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/SUCCESStrue"
                    },
                    {
                      "$ref": "#/components/schemas/EmailError.EMAIL_CERTIFICATION_SEND_LIMIT_EXCEEDED"
                    },
                    {
                      "$ref": "#/components/schemas/UserError.EMAIL_ALREADY_EXISTS"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "이메일 인증번호 발송 제한 초과",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EmailError.EMAIL_CERTIFICATION_SEND_LIMIT_EXCEEDED"
                }
              }
            }
          }
        },
        "summary": "이메일 인증번호 발송 API",
        "description": "이메일 인증번호 발송 API\n\n로컬 회원가입을 위한 이메일 인증번호를 발송합니다.\n이미 가입된 이메일인 경우 에러를 반환합니다.\n\n하루에 5회까지만 인증번호를 발송할 수 있습니다.\n인증번호는 15분간 유효합니다.\n\n15분이 지난 인증번호는 자동 만료됩니다."
      }
    },
    "/api/auth/email/certification/verify": {
      "post": {
        "tags": ["Auth"],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Auth.VerifyEmailCodeDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "$ref": "#/components/schemas/SUCCESS__type.o4"
                    },
                    {
                      "$ref": "#/components/schemas/UserError.EMAIL_ALREADY_EXISTS"
                    },
                    {
                      "$ref": "#/components/schemas/EmailError.EMAIL_CERTIFICATION_CODE_NOT_MATCH"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "이메일이 이미 존재합니다.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthError.User.EMAIL_ALREADY_EXISTS"
                }
              }
            }
          }
        },
        "summary": "이메일 인증번호 확인 API",
        "description": "이메일 인증번호 확인 API\n\n로컬 회원가입을 위한 이메일 인증번호를 확인합니다.\n인증번호가 일치하지 않으면 에러를 반환합니다.\n이미 가입된 이메일인 경우 에러를 반환합니다.\n인증번호는 15분간 유효합니다.\n\n인증완료시 인증번호ID를 반환합니다.\n해당 ID는 회원가입시 사용됩니다."
      }
    },
    "/api/auth/logout": {
      "post": {
        "tags": ["Auth"],
        "parameters": [
          {
            "name": "X-Client-Type",
            "in": "header",
            "schema": {
              "type": "string",
              "default": "'WEB'",
              "description": "Client type"
            },
            "description": "Client type",
            "required": false
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SUCCESS__type.o5"
                }
              }
            }
          }
        },
        "summary": "로그아웃 API",
        "description": "로그아웃 API\n\n로그아웃시, 리프레시 토큰을 만료시킵니다.\n\n**주의**\n현재 access_token에대한 blacklist를 사용하지 않습니다. (추후 변경될수 있습니다.)\n- access_token까지 redis로 관리할시, 세션과 사용하는 것과 같다 생각하여 사용하지 않습니다.\n\n따라서, 각 클라이언트는 로그아웃시 access_token을 삭제해야합니다.\n웹의 경우 access_token을 삭제해야하며,\n앱의 경우 access_token과 refresh_token을 삭제해야합니다."
      }
    }
  },
  "components": {
    "schemas": {
...
      "AuthError.Authentication.TOKEN_INVALID": {
        "type": "object",
        "properties": {
          "is_success": {
            "type": "boolean",
            "enum": [false]
          },
          "message": {
            "type": "string",
            "enum": ["TOKEN_INVALID"]
          },
          "status": {
            "type": "number",
            "enum": [401]
          }
        },
        "nullable": false,
        "required": ["is_success", "message", "status"]
      },
      "SUCCESS__type": {
        "type": "object",
        "properties": {
          "is_success": {
            "type": "boolean",
            "enum": [true],
            "description": "Whether the request was successful"
          },
          "message": {
            "type": "string",
            "description": "Message"
          },
          "result": {
            "type": "object",
            "properties": {
              "is_login": {
                "type": "boolean"
              },
              "user": {
                "$ref": "#/components/schemas/JwtPayload"
              }
            },
            "nullable": false,
            "required": ["is_login"],
            "description": "Result"
          }
        },
        "nullable": false,
        "required": ["is_success", "message", "result"]
      },
...
      "AuthError.Authentication.TOKEN_MISSING": {
        "type": "object",
        "properties": {
          "is_success": {
            "type": "boolean",
            "enum": [false]
          },
          "message": {
            "type": "string",
            "enum": ["TOKEN_MISSING"]
          },
          "status": {
            "type": "number",
            "enum": [400]
          }
        },
        "nullable": false,
        "required": ["is_success", "message", "status"]
      },
      "SUCCESSaccess_tokenstringrefresh_tokenstringaccess_tokenstring": {
        "type": "object",
        "properties": {
          "is_success": {
            "type": "boolean",
            "enum": [true],
            "description": "Whether the request was successful"
          },
          "message": {
            "type": "string",
            "description": "Message"
          },
          "result": {
            "oneOf": [
              {
                "type": "object",
                "properties": {
                  "access_token": {
                    "type": "string"
                  },
                  "refresh_token": {
                    "type": "string"
                  }
                },
                "nullable": false,
                "required": ["access_token", "refresh_token"]
              },
              {
                "type": "object",
                "properties": {
                  "access_token": {
                    "type": "string"
                  }
                },
                "nullable": false,
                "required": ["access_token"]
              }
            ],
            "description": "Result"
          }
        },
        "nullable": false,
        "required": ["is_success", "message", "result"]
      },
      "Auth.Oauth": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["google", "kakao", "naver"],
            "description": "로그인 타입\n- local: 이메일, 비밀번호\n- google: 구글 로그인\n- kakao: 카카오 로그인"
          },
          "code": {
            "type": "string",
            "description": "인증 코드"
          },
          "redirect_uri": {
            "type": "string",
            "description": "리다이렉트 URI"
          },
          "state": {
            "type": "string",
            "description": "state\nex:\nnaver"
          },
          "auto_login": {
            "default": true,
            "type": "boolean",
            "description": "자동 로그인 여부"
          }
        },
        "nullable": false,
        "required": ["type", "code", "redirect_uri"]
      },
      "Auth.Local": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "enum": ["local"],
            "description": "로그인 타입\n- local: 이메일, 비밀번호\n- google: 구글 로그인\n- kakao: 카카오 로그인"
          },
          "email": {
            "type": "string",
            "format": "email",
            "description": "이메일"
          },
          "password": {
            "type": "string",
            "minLength": 8,
            "description": "비밀번호"
          },
          "auto_login": {
            "default": true,
            "type": "boolean",
            "description": "자동 로그인 여부"
          }
        },
        "nullable": false,
        "required": ["type", "email", "password"]
      },
      "AuthError.Authentication.INVALID_TYPE": {
        "type": "object",
        "properties": {
          "is_success": {
            "type": "boolean",
            "enum": [false]
          },
          "message": {
            "type": "string",
            "enum": ["INVALID_TYPE"]
          },
          "status": {
            "type": "number",
            "enum": [400]
          }
        },
        "nullable": false,
        "required": ["is_success", "message", "status"]
      },
...
      "AuthError.User.USER_NOT_FOUND": {
        "type": "object",
        "properties": {
          "is_success": {
            "type": "boolean",
            "enum": [false]
          },
          "message": {
            "type": "string",
            "enum": ["USER_NOT_FOUND"]
          },
          "status": {
            "type": "number",
            "enum": [404]
          }
        },
        "nullable": false,
        "required": ["is_success", "message", "status"]
      },
      "SUCCESS__type.o2": {
        "type": "object",
        "properties": {
          "is_success": {
            "type": "boolean",
            "enum": [true],
            "description": "Whether the request was successful"
          },
          "message": {
            "type": "string",
            "description": "Message"
          },
          "result": {
            "type": "object",
            "properties": {
              "access_token": {
                "type": "string"
              }
            },
            "nullable": false,
            "required": ["access_token"],
            "description": "Result"
          }
        },
        "nullable": false,
        "required": ["is_success", "message", "result"]
      },
      "SUCCESS__type.o1": {
        "type": "object",
        "properties": {
          "is_success": {
            "type": "boolean",
            "enum": [true],
            "description": "Whether the request was successful"
          },
          "message": {
            "type": "string",
            "description": "Message"
          },
          "result": {
            "type": "object",
            "properties": {
              "access_token": {
                "type": "string"
              },
              "refresh_token": {
                "type": "string"
              }
            },
            "nullable": false,
            "required": ["access_token", "refresh_token"],
            "description": "Result"
          }
        },
        "nullable": false,
        "required": ["is_success", "message", "result"]
      },
...
    },
    "securitySchemes": {
      "web_refresh_token": {
        "type": "apiKey",
        "name": "_refresh_token",
        "in": "cookie"
      },
      "refresh_token": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      },
      "access_token": {
        "type": "apiKey",
        "name": "Authorization",
        "in": "header"
      }
    }
  }
}